Feature: introduced suspend/resume on rMem.
git-svn-id: http://cadia.ru.is/svn/repos/replicode@186 36b273a1-5dce-4f9d-8980-b90fca7d7570

+++ r_exec/view.cpp
@@ -213,7 +213,7 @@ namespace	r_exec{
 
 	NotificationView::NotificationView(Group	*origin,Group	*destination,Code	*marker):View(){
 
-		uint32	write_index=0;
+		uint16	write_index=0;
 		code(write_index++)=r_code::Atom::SSet(ViewOpcode,5);	//	Structured Set.
 		code(write_index++)=r_code::Atom::IPointer(6);			//	iptr to ijt.
 		code(write_index++)=r_code::Atom::Float(1);				//	sln.
